pyproject.toml
[tool.poetry]
name = "track"
version = "0.1.0"
description = ""
authors = ["Ulisse Mini <ulisse.mini@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.71.0"
uvicorn = {extras = ["standard"], version = "^0.17.0"}
dominate = "^2.6.0"
python-multipart = "^0.0.5"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
